+++ ProjectC/reservation/src/main/java/com/example/reservation/domain/ReservationUserComment.java
@@ -9,11 +9,11 @@ import java.util.Date;
 @Getter @Setter
 @NoArgsConstructor
 public class ReservationUserComment {
-    private int id;
-    private int product_id;
-    private int reservation_info_id;
-    private int user_id;
-    private int score;
+    private long id;
+    private long product_id;
+    private long reservation_info_id;
+    private long user_id;
+    private double score;
     private String comment;
     private Date create_date;
     private Date modify_date;
